Optimize the pre-commit docker image
The size goes from 1.36GB to 240MB.

First lets use a smaller base image. I've removed build-essential
because it seems to be not needed. Finally I've removed the virtual env
to simplify this image because we only have 1 Python environment.
apt-get clean is run automatically, but we do need to clean the apt-get
update lists.